tax collector last fix

+++ src/koh/game/conditions/EmoteCriterion.java
@@ -19,12 +19,11 @@ public class EmoteCriterion extends Criterion {
 
     @Override
     public void Build() {
-        this.emote = Byte.parseByte(literal);;
+        this.emote = (byte) Integer.parseInt(literal);;
     }
 
     @Override
     public boolean eval(Player character) {
-        //FIXME : See if another operator exist for this
         switch (this.operator) {
             case INEQUALS:
                 return !Bytes.contains(character.getEmotes(), emote);
